Editor: press and hold mechanic added, window made unresizeable, view movement slowed down for accuracy, UI now draws on top, maplength mechanic added, width and height changed to float. Game: maplength compatibility added, width and height changed to float.

+++ Mario/Editor/Resources.h
@@ -8,8 +8,8 @@
 // You should do it someday... :D
 
 // Contains the viewport sizes, but this should not ever change
-constexpr int WIDTH = 1920;
-constexpr int HEIGHT = 1080;
+constexpr float WIDTH = 1920;
+constexpr float HEIGHT = 1080;
 
 class Resources {
 private:
